# swagger-export.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ai/providers": {
      "get": {
        "operationId": "AIKeysController_getProviders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of AI providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AIProviderResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all AI providers",
        "tags": [
          "🤖 AI"
        ]
      }
    },
    "/ai/providers/enum": {
      "get": {
        "operationId": "AIController_getProviderEnum",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All possible AI provider values"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get AI provider enum values",
        "tags": [
          "🤖 AI"
        ]
      }
    },
    "/ai/keys": {
      "get": {
        "operationId": "AIKeysController_getUserKeys",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of user AI keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AIKeyResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user AI keys",
        "tags": [
          "🤖 AI"
        ]
      },
      "post": {
        "operationId": "AIKeysController_createUserKey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAIKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "AI key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIKeyResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new AI key",
        "tags": [
          "🤖 AI"
        ]
      }
    },
    "/ai/keys/{id}": {
      "put": {
        "operationId": "AIKeysController_updateUserKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "AI key ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAIKeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI key updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIKeyResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update AI key",
        "tags": [
          "🤖 AI"
        ]
      },
      "delete": {
        "operationId": "AIKeysController_deleteUserKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "AI key ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "AI key deleted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete AI key",
        "tags": [
          "🤖 AI"
        ]
      }
    },
    "/ai/admin/system-keys": {
      "get": {
        "operationId": "AIKeysController_getSystemKeys",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of system AI keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AIKeyResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all system AI keys (Admin only)",
        "tags": [
          "🤖 AI"
        ]
      },
      "post": {
        "operationId": "AIKeysController_createSystemKey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAIKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "System AI key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIKeyResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create system AI key (Admin only)",
        "tags": [
          "🤖 AI"
        ]
      }
    },
    "/ai/admin/system-keys/{id}": {
      "put": {
        "operationId": "AIKeysController_updateSystemKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "System AI key ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAIKeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "System AI key updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIKeyResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update system AI key (Admin only)",
        "tags": [
          "🤖 AI"
        ]
      },
      "delete": {
        "operationId": "AIKeysController_deleteSystemKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "System AI key ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "System AI key deleted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete system AI key (Admin only)",
        "tags": [
          "🤖 AI"
        ]
      }
    },
...
  },
  "info": {
    "title": "TechVerse Café API",
    "description": "Backend API for TechVerse Café MVP - A tech community platform with digital economy\n\n## 🚀 New Features in MVP\n- � **Digital Wallet & TechCoin System**: Earn, buy, and spend TechCoin\n- � **Challenges & Competitions**: Community challenges with rewards\n- �️ **Enhanced Projects System**: Task management with payments\n- 🌍 **Multi-language Support**: Localized content and UI\n- ☕ **Café Communities**: Topic-based discussion groups\n\n## Core Features\n- � JWT Authentication & Role-based Access Control\n- 📝 Content Management (Articles, Posts)\n- 🔖 Bookmarking System\n- 👥 User Profiles & Social Features\n- 🤖 AI Integration for content generation\n\n## Authentication\nMost endpoints require authentication using JWT tokens. Include the token in the `X-HTTP-TOKEN` header.\n\n## Role-Based Access Control\nThe platform uses a two-tier role system:\n- **Global Roles**: ADMIN, SUPERVISOR, EDITOR, MEMBER\n- **Café Roles**: BARISTA, THINKER, JOURNALIST, MEMBER\n\n## Getting Started\n1. Register a new user account\n2. Login to receive JWT token\n3. Use the token in `X-HTTP-TOKEN` header for authenticated requests\n4. Create your wallet and start earning TechCoin\n5. Join challenges and participate in projects\n\n## Download API Specification\n- JSON Format: [/api/swagger.json](/api/swagger.json)\n- YAML Format: [/api/swagger.yaml](/api/swagger.yaml)\n      ",
    "version": "2.0.0",
...
    },
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "X-HTTP-TOKEN": {
        "type": "apiKey",
        "in": "header",
        "name": "X-HTTP-TOKEN",
        "description": "JWT token for authentication. Format: Bearer <token>"
      }
    },
    "schemas": {
...
      "AIProviderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Provider ID"
          },
          "name": {
            "type": "string",
            "description": "Provider name (slug)"
          },
          "label": {
            "type": "string",
            "description": "Provider display label"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether provider is active"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Whether this is the default provider"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Created timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Updated timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "label",
          "isActive",
          "isDefault",
          "createdAt",
          "updatedAt"
        ]
      },
      "AIKeyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "API key ID"
          },
          "providerId": {
            "type": "number",
            "description": "Provider ID"
          },
          "keyName": {
            "type": "string",
            "description": "Custom key name"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether key is active"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Created timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Updated timestamp"
          },
          "provider": {
            "type": "object",
            "description": "Provider information"
          }
        },
        "required": [
          "id",
          "providerId",
          "keyName",
          "isActive",
          "createdAt",
          "updatedAt",
          "provider"
        ]
      },
      "CreateAIKeyDto": {
        "type": "object",
        "properties": {
          "providerId": {
            "type": "number",
            "description": "ID of the AI provider",
            "example": 1
          },
          "secretKey": {
            "type": "string",
            "description": "The API key secret",
            "example": "sk-..."
          }
        },
        "required": [
          "providerId",
          "secretKey"
        ]
      },
      "UpdateAIKeyDto": {
        "type": "object",
        "properties": {
          "secretKey": {
            "type": "string",
            "description": "Updated API key secret",
            "example": "sk-..."
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the API key is active",
            "example": true
          }
        }
      },
...
    }
  }
}